+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 ### Added
+- Can now be called from command line to start OAuth2 wizards.
 - Sending emails also supports a tuple of content and filename for attachments.
 
 ### Changed